+++ src/module/sheets/AbstractTwodsixActorSheet.ts
@@ -9,7 +9,6 @@ import {onPasteStripFormatting} from "../sheets/AbstractTwodsixItemSheet";
 //import { getKeyByValue } from "../utils/sheetUtils";
 import { resolveUnknownAutoMode } from "../utils/rollItemMacro";
 import { TWODSIX } from "../config";
-//import { applyEncumberedEffect } from "../hooks/showStatusIcons";
 
 export abstract class AbstractTwodsixActorSheet extends ActorSheet {
 
@@ -60,7 +59,7 @@ export abstract class AbstractTwodsixActorSheet extends ActorSheet {
           yes: async () => {
             const selectedActor = this.actor.isToken ? this.token?.actor : this.actor;
             await ownedItem.update({'system.equipped': 'ship'});
-            await selectedActor?.deleteEmbeddedDocuments("Item", [<string>ownedItem.id]);
+            await selectedActor?.deleteEmbeddedDocuments("Item", [ownedItem.id]);
             // somehow on hooks isn't working when a consumable is deleted  - force the issue
             if (ownedItem.type === "consumable") {
               selectedActor?.items.filter(i => i.type !== "skills" && i.type !== "trait").forEach(async i => {
@@ -577,8 +576,8 @@ export abstract class AbstractTwodsixActorSheet extends ActorSheet {
           const itemWithEffect = await fromUuid(selectedEffect.origin);
           await itemWithEffect?.update({effects: [] }, {recursive: false});  //can't directly delete using deleteEmbeddedDocuments
         }
-        await selectedEffect?.delete();
-        //await this.actor.deleteEmbeddedDocuments('ActiveEffect', [selectedEffect.id]);
+        //await selectedEffect?.delete();
+        await this.actor.deleteEmbeddedDocuments('ActiveEffect', [selectedEffect?.id]);
       },
       no: () => {
         //Nothing